refactored project to mySQL database with work on hibernate framework

+++ src/main/java/com/game/controller/PlayerController.java
@@ -38,7 +38,7 @@ public class PlayerController {
     }
 
     @GetMapping("/count")
-    public Integer getAllCount() {
+    public Long getAllCount() {
         return playerService.getAllCount();
     }
